gateway/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
requests==2.26.0
Pillow==10.3.0
prometheus_client==0.14.1
